Work on new scheduler

+++ hapi-fhir-jpaserver-base/src/main/java/ca/uhn/fhir/jpa/util/TestUtil.java
@@ -148,7 +148,7 @@ public class TestUtil {
 		}
 	}
 
-	public static void sleepAtLeast(int theMillis) {
+	public static void sleepAtLeast(long theMillis) {
 		long start = System.currentTimeMillis();
 		while (System.currentTimeMillis() <= start + theMillis) {
 			try {
